Ricochet Bullet
It took a lot of messing around but the Ricochet bullet works now. However, it seems like the physics in Godot works in a way that makes bullets ricochet 1 frame later than they should. Because the bullets are moving so fast, they often turn around after completely passing through an object and can even collide with the same object again afterwards.

- Added a Raycast2D child to both bullet node scenes which is used to get the needed ricochet angle
- Reorganized nodes in starting room scene so all enemies are children of the Enemies node
- Added a "ricochets" export variable for bullets that tells how many times they can ricochet