Basic Pickups
Added a new bullet pickup scene that automatically gives you ammo if you have the corresponding bullet type in your inventory. Also added a signal to enemy script (sorry Nick) so that bullets pickups can be instantiated in the level at the location that enemies are killed whenever they die.
- WIP animation player for HUD make rotation based on time rather than frames so that higher framerate doesn't make the game easier