Railgun + Explosive Bullets
Made a new player bullet script that extends the basic bullet script so that I didn't clog it up too much for the enemy bullet variant. Explosive and Railgun bullets now have all of their intended effects (but not finalized visuals). Also rearranged some files so that all bullet-related scenes/scripts are in the same folder